refactor: globalization of collapsible-content
